refactor: extract routers/contexts.py; main.py becomes app setup only
Move all remaining context routes (/, /ui/*, /api/contexts/*) out of
main.py into api/routers/contexts.py. main.py is now app setup,
lifespan, middleware, and router includes (~76 lines).
